refactor: ActionResult 型の移動 + auth 経路統一 + デッドコード削除 (Epic #288 PBI-3)
- R3: ActionResult<T> と extractFieldErrors を src/lib/validations/materials.ts から
  src/lib/types/action-result.ts に移動。教材バリデーションと無関係な 7+ ファイル
  が materials.ts から共通型を import する責務混在を解消。
  validations/sessions.ts と validations/notifications.ts の迂回 re-export
  (export { type ActionResult, extractFieldErrors } from "./materials") も削除し、
  呼び出し元が直接 @/lib/types/action-result を import するよう統一。

- R5: src/app/(main)/profile/page.tsx と profile/actions.ts を requireAuth()
  経由に統一。profile は認証必須のため user の non-null を型で保証する。
  (main)/layout.tsx は middleware による認証保護を前提に未認証を意図的に許容
  している旨をコメントで明示。

- デッドコード削除: Epic #232 のマージ後残置エイリアスを削除:
  - createSubject, getSubjects (src/lib/actions/categories.ts)
  - createSubjectSchema, CreateSubjectInput (src/lib/validations/materials.ts)
  - Subject (src/lib/types/materials.ts)
  呼び出し元は tests/small/lib/validations/materials.test.ts のみで
  createCategorySchema に書き換え済み。

tests/shared/helpers.ts の createTestSubject は 47 箇所で使用中につき削除対象外。
テスト記述量削減のためエイリアスとして維持する意図をコメントで明示。

closes #291
