+++ tests/shared/helpers.ts
@@ -14,7 +14,8 @@ export async function createTestCategory(
   return result.data as { id: string; name: string; color: string; user_id: string; parent_id: string | null };
 }
 
-// 後方互換エイリアス。Epic #232 全 PBI マージ完了時に削除予定 (呼び出し元 20+ 箇所)
+// createTestCategory の短縮エイリアス。テスト記述量削減のためテストヘルパーとしてのみ維持する
+// (Epic #232 のデッドコード削除は Epic #288 PBI-3 で完了済み。本エイリアスは削除対象外)
 export const createTestSubject = createTestCategory;
 
 export async function createTestMaterial(
